feat(UIButtonBuilder): expose cached index operations
The CachedIndex can now be accessed via the `TryGetCachedIndex` method
which will return the current cached index for the selected item if
one has been set, otherwise the method will return false.

The `SwitchTo(int index)` method has also been exposed publicly so
switching can happen programatically on the component and the relevant
index cache is updated accordingly.

The `SwitchToCachedIndex` method also now has a
`defaultWhenNoCachedIndex` optional parameter which can be used as the
index to actually switch to if no cached index is present.

The prefab has been updated to use this new `SwitchToCachedIndex`
method when the prefab is enabled.